split metadata extraction to separate file
